+++ src/widgets/trackslider.cpp
@@ -185,7 +185,7 @@ void TrackSlider::ValueMaybeChanged(const int value) {
   if (setting_value_) return;
 
   UpdateTimes(static_cast<int>(value / kMsecPerSec));
-  emit ValueChangedSeconds(static_cast<int>(value / kMsecPerSec));
+  emit ValueChangedSeconds(static_cast<quint64>(value / kMsecPerSec));
 
 }
